pyproject.toml
[project]
name = "umu-sales-trainer"
version = "0.1.0"
description = "UMU Sales Trainer - AI-powered sales training system"
requires-python = ">=3.13"
dependencies = [
    "langgraph>=1.0.0",
    "langchain-core>=0.3.0",
    "langchain-openai>=0.2.0",
    "openai>=1.50.0",
    "dashscope>=1.20.0",
    "sentence-transformers>=3.0.0",
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.30.0",
    "pyyaml>=6.0.0",
    "pydantic>=2.0.0",
    "pydantic-settings>=2.0.0",
    "aiosqlite>=0.20.0",
    "chromadb>=0.4.0",
    "langchain-chroma>=0.1.0",
    "httpx>=0.27.0",
    "sqlalchemy>=2.0.49",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.24.0",
    "pytest-cov>=6.0.0",
    "ruff>=0.8.0",
    "basedpyright>=1.0.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/umu_sales_trainer"]

[tool.ruff]
target-version = "py313"
line-length = 100

[tool.basedpyright]
pythonVersion = "3.13"

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[tool.coverage.run]
source = ["src"]
branch = true

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise AssertionError",
    "raise NotImplementedError",
    "if __name__ == .__main__.:",
]

[dependency-groups]
dev = [
    "basedpyright>=1.39.0",
    "pytest>=9.0.2",
    "pytest-asyncio>=1.3.0",
    "pytest-cov>=7.1.0",
    "pytest-mock>=3.15.1",
    "ruff>=0.15.9",
]
